+++ custom_components/silla_prism/sensor.py
@@ -356,6 +356,8 @@ class PrismSolarBalanceSensor(SensorEntity):
             self._attr_native_value = state.theoretical_target_current
         elif self.entity_description.value_key == "battery_reserve_power":
             self._attr_native_value = state.battery_reserve_power
+        elif self.entity_description.value_key == "battery_reserve_shortfall_power":
+            self._attr_native_value = state.battery_reserve_shortfall_power
         elif self.entity_description.value_key == "target_export_power":
             self._attr_native_value = state.target_export_power
         elif self.entity_description.value_key == "unused_export_power":
@@ -383,6 +385,7 @@ class PrismSolarBalanceSensor(SensorEntity):
             "battery_max_charge_power": state.battery_max_charge_power,
             "battery_soc": state.battery_soc,
             "battery_reserve_power": state.battery_reserve_power,
+            "battery_reserve_shortfall_power": state.battery_reserve_shortfall_power,
             "surplus_source": state.surplus_source,
             "target_export_power": state.target_export_power,
             "deadband_power": state.deadband_power,
@@ -529,6 +532,16 @@ SOLAR_BALANCE_SENSORS: tuple[PrismSolarBalanceSensorEntityDescription, ...] = (
         has_entity_name=True,
         translation_key="solar_balance_battery_reserve_power",
         value_key="battery_reserve_power",
+    ),
+    PrismSolarBalanceSensorEntityDescription(
+        key="solar_balance_battery_reserve_shortfall_power_{}",
+        device_class=SensorDeviceClass.POWER,
+        state_class=SensorStateClass.MEASUREMENT,
+        native_unit_of_measurement=UnitOfPower.WATT,
+        suggested_display_precision=0,
+        has_entity_name=True,
+        translation_key="solar_balance_battery_reserve_shortfall_power",
+        value_key="battery_reserve_shortfall_power",
     ),
     PrismSolarBalanceSensorEntityDescription(
         key="solar_balance_target_export_power_{}",